File Handling using Automation Selenium
In this project, we are given username and password of different users in excel and xml files. Automate the scenario of 'LogIn' and 'LogOut' using data (username and password) from the files in a browser. This helps to fetch data from the files so that we can log in with multiple users in an app. 
This is done using Selenium with Java. One can use any app to automate and try this scenario.